+++ Dia/DiaCore/Containers/Arrays/DynamicArray.inl
@@ -498,7 +498,7 @@ namespace Dia
 				{
 					T possibleUnique = At(i);
 					bool foundInList = false;
-					for (size_t j = 0; j < unique.Size(); j++)
+					for (unsigned int j = 0; j < unique.Size(); j++)
 					{
 						if (possibleUnique == unique[j])
 						{
@@ -522,7 +522,7 @@ namespace Dia
 				{
 					T possibleUnique = At(i);
 					bool foundInList = false;
-					for (size_t j = 0; j < unique.Size(); j++)
+					for (unsigned int j = 0; j < unique.Size(); j++)
 					{
 						if (possibleUnique == unique[j])
 						{
