+++ Dia/DiaCore/Containers/HashTables/HashTable.inl
@@ -96,7 +96,7 @@ namespace Dia
 			unsigned int HashTable<Key, Payload, HashFunctor >::DeepestTable() const
 			{
 				unsigned int deepest = 0;
-				for (size_t i = 0; i < mTable.Size(); i++)
+				for (unsigned int i = 0; i < mTable.Size(); i++)
 				{
 					if (mTable[i])
 					{
@@ -119,7 +119,7 @@ namespace Dia
 			unsigned int HashTable<Key, Payload, HashFunctor >::AssignedTableEntries() const
 			{
 				unsigned int numberAssignedTableEntries = 0;
-				for (size_t i = 0; i < mTable.Size(); i++)
+				for (unsigned int i = 0; i < mTable.Size(); i++)
 				{
 					if (mTable[i])
 					{
@@ -189,7 +189,7 @@ namespace Dia
 			void HashTable<Key, Payload, HashFunctor >::Remove ( const Key& key )
 			{
 				DIA_ASSERT_SUPPORT(bool found = false);
-				for (size_t i = 0; i < mPayloadNodes.Size(); i++)
+				for (unsigned int i = 0; i < mPayloadNodes.Size(); i++)
 				{
 					if (mPayloadNodes[i].GetKeyConst() == key)
 					{
@@ -232,7 +232,7 @@ namespace Dia
 			void HashTable<Key, Payload, HashFunctor >::RemoveByPayload ( const Payload& value )
 			{
 				DIA_ASSERT_SUPPORT(bool found = false);
-				for (size_t i = 0; i < mPayloadNodes.Size(); i++)
+				for (unsigned int i = 0; i < mPayloadNodes.Size(); i++)
 				{
 					if (mPayloadNodes[i].GetPayloadConst() == value)
 					{
@@ -249,7 +249,7 @@ namespace Dia
 			void HashTable<Key, Payload, HashFunctor >::RemoveByPayload( const Payload& payloadValue, const Comparisson& functor )
 			{
 				DIA_ASSERT_SUPPORT(bool found = false);
-				for (size_t i = 0; i < mPayloadNodes.Size(); i++)
+				for (unsigned int i = 0; i < mPayloadNodes.Size(); i++)
 				{
 					if (functor.Equals(mPayloadNodes[i].GetPayloadConst(), value))
 					{
@@ -265,7 +265,7 @@ namespace Dia
 			template< class Key, class Payload, class HashFunctor > 
 			void HashTable<Key, Payload, HashFunctor >::RemoveAll()
 			{
-				for(size_t i = 0; i < mTable.Size(); i++)
+				for(unsigned int i = 0; i < mTable.Size(); i++)
 				{
 					mTable[i] = NULL;
 				}
@@ -285,7 +285,7 @@ namespace Dia
 			template< class Key, class Payload, class HashFunctor > 
 			bool HashTable<Key, Payload, HashFunctor >::ContainsPayload ( const Payload& value ) const
 			{
-				for (size_t i = 0; i < mPayloadNodes.Size(); i++)
+				for (unsigned int i = 0; i < mPayloadNodes.Size(); i++)
 				{
 					if (mPayloadNodes[i].GetPayloadConst() == value)
 					{
@@ -299,7 +299,7 @@ namespace Dia
 			template< class Key, class Payload, class HashFunctor > template<class Equal> 
 			bool HashTable<Key, Payload, HashFunctor >::ContainsPayload	( const Payload& value, const Equal& functor ) const
 			{
-				for (size_t i = 0; i < mPayloadNodes.Size(); i++)
+				for (unsigned int i = 0; i < mPayloadNodes.Size(); i++)
 				{
 					if (functor.Equals(mPayloadNodes[i].GetPayloadConst(), value))
 					{
